Document the render() and status() methods
Both methods completely lacked any documentation explaining what they do. To fix this, documentation was added describing what the code tries to achieve, and how it goes about achieving it. This commit makes it much easier to understand and visualize the code, especially since both methods are used to display text and images to a screen.