Revise Documentation
This small commit improves one comment and removes another. A more accurate description of the render method is provided: "This method displays the current state of the game". Moreover, one redundant comment is removed, "Keep looping until game ends", since it is evident that this is what the code does: "while(Game.isRunning())".